Change href to React Link and rename Link Component
