+++ src/app/components/LinkComponent/LinkComponent.stories.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+import LinkComponent from "./LinkComponent";
+
+export default {
+  title: "Component/LinkComponent",
+  component: LinkComponent,
+};
+
+export const RegisterDog = (): JSX.Element => (
+  <LinkComponent link="#" variant="primary">
+    Hund registrieren
+  </LinkComponent>
+);
+
+export const RegisterSitter = (): JSX.Element => (
+  <LinkComponent link="#" variant="primary">
+    Hundesitter werden
+  </LinkComponent>
+);
+
+export const LogIn = (): JSX.Element => (
+  <LinkComponent link="#" variant="secondary">
+    Log-In
+  </LinkComponent>
+);
